requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
openai==1.3.0
python-dotenv==1.0.0
pyaudio==0.2.13
google-cloud-speech==2.20.0
google-cloud-texttospeech==2.13.0
requests==2.31.0
flask==2.3.0
pynput==1.7.6
psutil==5.9.5
